Will start adding nan-resistance now.
The idea is as follows: if there's a NaN in the input,
simply ignore the weights directly attached to the input
when computing the gradient. In order to do this, we'll
add a binary 'nograd' tensor of the same shape as the
first hidden layer to the net. The weight matrix will
be multiplied by this on each iteration.

The NaNs will be switched to an arbitrary value (0 in the
implementation) that won't be used in the computation,
as the nograd flag will be set to 0 when NaNs are present.